Refactor transition context and improve code consistency
- Moved transitionStore from TransitionContext to a new file TransitionStore.ts for better modularity.
- Updated imports in LinkWait, ScreenTransition, and constants to reflect the new TransitionStore location.
- Enhanced formatting and consistency in NavDock, Dock, and other components by standardizing quotes and spacing.
- Improved mobile detection logic in useIsMobile hook for better readability.
- Minor CSS adjustments for improved styling consistency.
- Updated Vite configuration to switch quote style from single to double for consistency across the project.

+++ vite.config.ts
@@ -9,7 +9,7 @@ export default defineConfig({
     tanstackRouter({
       autoCodeSplitting: true,
       generatedRouteTree: "src/routeTree.gen.ts",
-      quoteStyle: "single",
+      quoteStyle: "double",
       routeFileIgnorePrefix: "-",
       routesDirectory: "src/routes",
       target: "react",
@@ -17,7 +17,7 @@ export default defineConfig({
     tailwindcss(),
     react({
       babel: {
-        plugins: ['babel-plugin-react-compiler'],
+        plugins: ["babel-plugin-react-compiler"],
       },
     }),
   ],
